modules/YoloModule/app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.12.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.5.25=hecd8cb5_1
cachetools=4.2.2=pypi_0
cairo=1.14.12=hc4e6be7_4
certifi=2020.12.5=py37hecd8cb5_0
chardet=4.0.0=pypi_0
click=8.0.1=pypi_0
cycler=0.10.0=py37_0
easydict=1.9=pypi_0
ffmpeg=4.0=h01ea3c9_0
flask=2.0.1=pypi_0
fontconfig=2.13.1=ha9ee91d_0
freetype=2.10.4=ha233b18_0
gast=0.3.3=pypi_0
gettext=0.21.0=h7535e17_0
glib=2.63.1=hd977a24_0
google-auth=1.30.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=h38d11af_0
grpcio=1.38.0=pypi_0
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
idna=2.10=pypi_0
importlib-metadata=4.3.1=pypi_0
intel-openmp=2021.2.0=hecd8cb5_564
itsdangerous=2.0.1=pypi_0
jasper=2.0.14=h636a363_1
jinja2=3.0.1=pypi_0
jpeg=9b=he5867d9_2
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37h23ab428_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libedit=3.1.20210216=h9ed2024_1
libffi=3.2.1=h0a44026_1007
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h23ab428_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py37hecd8cb5_0
matplotlib-base=3.3.4=py37h8b3ea08_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py37h9ed2024_1
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
ncurses=6.2=h0a44026_1
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=4.1.1.26=pypi_0
openssl=1.0.2u=h1de35cc_0
opt-einsum=3.3.0=pypi_0
owlready2=0.31=pypi_0
pcre=8.44=hb1e8313_0
pillow=8.2.0=py37h5270095_0
pip=21.1.1=py37hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
protobuf=3.17.1=pypi_0
py-opencv=3.4.2=py37h7c891bd_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.7.0=hc167b69_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
readline=7.0=h1de35cc_5
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.3.0rc0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=2.3.0.dev2020062301=pypi_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py37h9ed2024_0
tqdm=4.61.0=pypi_0
typing-extensions=3.10.0.0=pypi_0
urllib3=1.26.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
